Consolidate and expand runner security tests
- Merge and relocate validator tests into existing test files,
  removing duplicate test files.
- Add tests for directory and file permission validation, output
  directory access, shell/network detection, and dangerous command
  heuristics.
- Add custom-config and real-UID integration-style tests for validator.
- Apply pre-commit end-of-file fixes to test files.
