--- internal/runner/security/directory_permission_test.go
@@ -1,264 +1,0 @@
-//go:build test
-
-package security
-
-import (
-	"os"
-	"os/user"
-	"path/filepath"
-	"strconv"
-	"testing"
-
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
-
-	commontesting "github.com/isseis/go-safe-cmd-runner/internal/common/testing"
-	"github.com/isseis/go-safe-cmd-runner/internal/groupmembership"
-)
-
-func TestValidator_validateDirectoryComponentPermissions_WithRealUID(t *testing.T) {
-	// Get current user info for testing
-	currentUser, err := user.Current()
-	require.NoError(t, err)
-
-	currentUID, err := strconv.Atoi(currentUser.Uid)
-	require.NoError(t, err)
-
-	otherUID := currentUID + 1000 // Use a different UID for testing
-
-	tests := []struct {
-		name        string
-		setupFunc   func(mockFS *commontesting.MockFileSystem)
-		realUID     int
-		wantErr     bool
-		errContains string
-	}{
-		{
-			name: "owner_write_permission_with_matching_uid",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// Directory with owner write permission, owned by current user
-				currentGid, err := strconv.ParseUint(currentUser.Gid, 10, 32)
-				require.NoError(t, err)
-				err = mockFS.AddDirWithOwner("/test-dir", 0o755, uint32(currentUID), uint32(currentGid))
-				require.NoError(t, err)
-			},
-			realUID: currentUID,
-			wantErr: false,
-		},
-		{
-			name: "owner_write_permission_with_non_matching_uid",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// Directory with owner write permission, owned by different user
-				currentGid, err := strconv.ParseUint(currentUser.Gid, 10, 32)
-				require.NoError(t, err)
-				err = mockFS.AddDirWithOwner("/test-dir", 0o755, uint32(otherUID), uint32(currentGid))
-				require.NoError(t, err)
-			},
-			realUID:     currentUID,
-			wantErr:     true,
-			errContains: "is owned by UID",
-		},
-		{
-			name: "root_owned_directory_always_allowed",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// Root-owned directory should always be allowed
-				err := mockFS.AddDirWithOwner("/test-dir", 0o755, UIDRoot, GIDRoot)
-				require.NoError(t, err)
-			},
-			realUID: currentUID,
-			wantErr: false,
-		},
-		{
-			name: "group_write_permission_with_single_group_member",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// Directory with group write permission using current user's group
-				// Use permissive mode for this test to avoid environmental complexities
-				currentGid, err := strconv.ParseUint(currentUser.Gid, 10, 32)
-				require.NoError(t, err)
-				err = mockFS.AddDirWithOwner("/test-dir", 0o775, uint32(currentUID), uint32(currentGid))
-				require.NoError(t, err)
-			},
-			realUID: currentUID,
-			// This test should pass because testPermissiveMode is enabled
-			wantErr: false,
-		},
-		{
-			name: "world_writable_directory_rejected",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// World-writable directory should be rejected
-				currentGid, err := strconv.ParseUint(currentUser.Gid, 10, 32)
-				require.NoError(t, err)
-				err = mockFS.AddDirWithOwner("/test-dir", 0o777, uint32(currentUID), uint32(currentGid))
-				require.NoError(t, err)
-			},
-			realUID:     currentUID,
-			wantErr:     true,
-			errContains: "writable by others",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			// Create mock filesystem and validator
-			mockFS := commontesting.NewMockFileSystem()
-			tt.setupFunc(mockFS)
-
-			// Create actual group membership (will be used for real group checking)
-			groupMembership := groupmembership.New()
-
-			config := DefaultConfig()
-			// For the problematic test, use permissive mode to bypass group membership complexities
-			if tt.name == "group_write_permission_with_single_group_member" {
-				config.testPermissiveMode = true
-			}
-
-			validator, err := NewValidatorWithFSAndGroupMembership(config, mockFS, groupMembership)
-			require.NoError(t, err) // Get directory info
-			info, err := mockFS.Lstat("/test-dir")
-			require.NoError(t, err)
-
-			// Test the function
-			err = validator.validateDirectoryComponentPermissions("/test-dir", info, tt.realUID)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-				if tt.errContains != "" {
-					assert.Contains(t, err.Error(), tt.errContains)
-				}
-			} else {
-				assert.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestValidator_validateCompletePath(t *testing.T) {
-	currentUser, err := user.Current()
-	require.NoError(t, err)
-
-	currentUID, err := strconv.Atoi(currentUser.Uid)
-	require.NoError(t, err)
-
-	tests := []struct {
-		name      string
-		setupFunc func(mockFS *commontesting.MockFileSystem)
-		path      string
-		realUID   int
-		wantErr   bool
-	}{
-		{
-			name: "complete_path_validation_with_uid_context",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// Create a complete directory hierarchy owned by current user
-				err := mockFS.AddDirWithOwner("/home", 0o755, UIDRoot, GIDRoot)
-				require.NoError(t, err)
-
-				err = mockFS.AddDirWithOwner("/home/user", 0o755, uint32(currentUID), 1000)
-				require.NoError(t, err)
-
-				err = mockFS.AddDirWithOwner("/home/user/project", 0o755, uint32(currentUID), 1000)
-				require.NoError(t, err)
-			},
-			path:    "/home/user/project",
-			realUID: currentUID,
-			wantErr: false,
-		},
-		{
-			name: "complete_path_validation_with_ownership_mismatch",
-			setupFunc: func(mockFS *commontesting.MockFileSystem) {
-				// Create hierarchy with ownership mismatch
-				err := mockFS.AddDirWithOwner("/tmp", 0o755, UIDRoot, GIDRoot)
-				require.NoError(t, err)
-
-				err = mockFS.AddDirWithOwner("/tmp/other", 0o755, uint32(currentUID+1000), 1000) // Different owner
-				require.NoError(t, err)
-			},
-			path:    "/tmp/other",
-			realUID: currentUID,
-			wantErr: true,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			mockFS := commontesting.NewMockFileSystem()
-			tt.setupFunc(mockFS)
-
-			groupMembership := groupmembership.New()
-
-			config := DefaultConfig()
-			validator, err := NewValidatorWithFSAndGroupMembership(config, mockFS, groupMembership)
-			require.NoError(t, err)
-
-			cleanPath := filepath.Clean(tt.path)
-			err = validator.validateCompletePath(cleanPath, tt.path, tt.realUID)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-		})
-	}
-}
-
-func TestValidator_validateOutputDirectoryAccess_WithImprovedLogic(t *testing.T) {
-	currentUser, err := user.Current()
-	require.NoError(t, err)
-
-	currentUID, err := strconv.Atoi(currentUser.Uid)
-	require.NoError(t, err)
-
-	tests := []struct {
-		name      string
-		setupFunc func(t *testing.T) string
-		realUID   int
-		wantErr   bool
-	}{
-		{
-			name: "output_directory_owned_by_user",
-			setupFunc: func(t *testing.T) string {
-				tempDir := t.TempDir()
-
-				// Create subdirectory owned by current user
-				subDir := filepath.Join(tempDir, "output")
-				err := os.MkdirAll(subDir, 0o755)
-				require.NoError(t, err)
-
-				return subDir
-			},
-			realUID: currentUID,
-			wantErr: false,
-		},
-		{
-			name: "non_existent_directory_with_existing_parent",
-			setupFunc: func(t *testing.T) string {
-				tempDir := t.TempDir()
-
-				// Return path to non-existent subdirectory
-				return filepath.Join(tempDir, "nonexistent", "output")
-			},
-			realUID: currentUID,
-			wantErr: false, // Should succeed if parent is accessible
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			config := DefaultConfig()
-			config.testPermissiveMode = true // Use permissive mode for real filesystem tests
-
-			validator, err := NewValidatorWithGroupMembership(config, nil)
-			require.NoError(t, err)
-
-			dirPath := tt.setupFunc(t)
-			err = validator.validateOutputDirectoryAccess(dirPath, tt.realUID)
-
-			if tt.wantErr {
-				assert.Error(t, err)
-			} else {
-				assert.NoError(t, err)
-			}
-		})
-	}
-}
